EACQAPW-5765. Update `Receipt` model
