+++ TinkoffASDKCore/TinkoffASDKCore/Domain/AcquiringAPI/Models/FfdVersion.swift
@@ -4,10 +4,11 @@
 //
 //  Created by Никита Васильев on 02.08.2023.
 //
+// swiftlint:disable identifier_name
 
 import Foundation
 
 public enum FfdVersion: String, Codable {
-    case version105 = "1.05"
-    case version112 = "1.2"
+    case version1_05 = "1.05"
+    case version1_2 = "1.2"
 }
