+++ TinkoffASDKCore/TinkoffASDKCore/Domain/AcquiringAPI/Models/Receipt.swift
@@ -4,101 +4,22 @@
 //
 //  Created by r.akhmadeev on 07.10.2022.
 //
+// swiftlint:disable identifier_name
 
 import Foundation
 
-/// Информация о чеке
-public struct Receipt: Encodable {
-    private enum CodingKeys: String, CodingKey {
-        case shopCode = "ShopCode"
-        case email = "Email"
-        case taxation = "Taxation"
-        case phone = "Phone"
-        case items = "Items"
-        case agentData = "AgentData"
-        case supplierInfo = "SupplierInfo"
-        case customer = "Customer"
-        case customerInn = "CustomerInn"
-        case ffdVersion = "FfdVersion"
-    }
-
-    /// Код магазина
-    public var shopCode: String?
-    /// Электронный адрес для отправки чека покупателю
-    /// Параметр `email` или `phone` должен быть заполнен.
-    public var email: String?
-    /// Телефон покупателя
-    /// Параметр `email` или `phone` должен быть заполнен.
-    public var phone: String?
-    /// Система налогообложения
-    public var taxation: Taxation?
-    /// Массив, содержащий в себе информацию о товарах
-    public var items: [Item]?
-    /// Данные агента
-    public var agentData: AgentData?
-    /// Данные поставщика платежного агента
-    public var supplierInfo: SupplierInfo?
-    /// Идентификатор покупателя
-    public var customer: String?
-    /// Инн покупателя. Если ИНН иностранного гражданина, необходимо указать 00000000000
-    public var customerInn: String?
-    /// Версия ФФД.
-    public var ffdVersion: FfdVersion?
-
-    /// Инициализиатор может пробросить ошибку в случае передачи невалидных полей
-    ///
-    /// НСПК требует как минимимум 1 валидное поле `phone` или `email` для формирования чека
-    public init(
-        shopCode: String?,
-        email: String?,
-        taxation: Taxation?,
-        phone: String?,
-        items: [Item]?,
-        agentData: AgentData?,
-        supplierInfo: SupplierInfo?,
-        customer: String?,
-        customerInn: String?,
-        ffdVersion: FfdVersion?
-    ) {
-        self.shopCode = shopCode
-        self.email = email
-        self.taxation = taxation
-        self.phone = phone
-        self.items = items
-        self.agentData = agentData
-        self.supplierInfo = supplierInfo
-        self.customer = customer
-        self.customerInn = customerInn
-        self.ffdVersion = ffdVersion
-    }
+public enum Receipt: Equatable, Encodable {
+    case version1_05(ReceiptFdv1_05)
+    case version1_2(ReceiptFdv1_2)
 
     public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        if shopCode != nil { try? container.encode(shopCode, forKey: .shopCode) }
-        if taxation != nil { try? container.encode(taxation?.rawValue, forKey: .taxation) }
-        if email != nil { try? container.encode(email, forKey: .email) }
-        if phone != nil { try? container.encode(phone, forKey: .phone) }
-        if items != nil { try? container.encode(items, forKey: .items) }
-        if agentData != nil { try? container.encode(agentData, forKey: .agentData) }
-        if supplierInfo != nil { try? container.encode(supplierInfo, forKey: .supplierInfo) }
-        if customer != nil { try? container.encode(customer, forKey: .customer) }
-        if customerInn != nil { try? container.encode(customerInn, forKey: .customerInn) }
-        if ffdVersion != nil { try? container.encode(ffdVersion, forKey: .ffdVersion) }
-    }
-}
+        var container = encoder.singleValueContainer()
 
-extension Receipt: Equatable {
-    public static func == (lhs: Receipt, rhs: Receipt) -> Bool {
-        return
-            lhs.shopCode == rhs.shopCode &&
-            lhs.email == rhs.email &&
-            lhs.phone == rhs.phone &&
-            lhs.taxation == rhs.taxation &&
-            lhs.items == rhs.items &&
-            lhs.agentData == rhs.agentData &&
-            lhs.supplierInfo == rhs.supplierInfo &&
-            lhs.customer == rhs.customer &&
-            lhs.customerInn == rhs.customerInn &&
-            lhs.ffdVersion == rhs.ffdVersion
+        switch self {
+        case let .version1_05(info):
+            try container.encode(info)
+        case let .version1_2(info):
+            try container.encode(info)
+        }
     }
 }
