+++ TinkoffASDKUI/TinkoffASDKUITests/TestCases/SBPServices/SBPPaymentService/SBPPaymentServiceTests.swift
@@ -204,17 +204,16 @@ extension PaymentFlow {
 
         let receiptItems: [Item] = []
 
-        paymentData.receipt = Receipt(
-            shopCode: nil,
-            email: "email@email.com",
-            taxation: .osn,
-            phone: "+79876543210",
-            items: receiptItems,
-            agentData: nil,
-            supplierInfo: nil,
-            customer: nil,
-            customerInn: nil,
-            ffdVersion: .version105
+        paymentData.receipt = .version1_05(
+            ReceiptFdv1_05(
+                shopCode: nil,
+                email: "email@email.com",
+                taxation: .osn,
+                phone: "+79876543210",
+                items: receiptItems,
+                agentData: nil,
+                supplierInfo: nil
+            )
         )
 
         let paymentOptions = PaymentOptions.create(from: paymentData)
